+++ src/writers/sink.creator.hh
@@ -8,15 +8,23 @@
 
 #include <optional>
 #include <memory>
+#include <unordered_map>
 
 namespace acquire::sink::zarr {
-struct SinkCreator final
+class SinkCreator final
 {
   public:
     SinkCreator() = delete;
     SinkCreator(std::shared_ptr<common::ThreadPool> thread_pool_,
                 std::shared_ptr<common::S3ConnectionPool> connection_pool);
     ~SinkCreator() noexcept = default;
+
+    /// @brief Create a sink from a URI and a path.
+    /// @param[in] base_uri The base URI for the sink.
+    /// @param[in] path The path for the sink.
+    /// @return The sink created, or nullptr if the URI is invalid.
+    std::unique_ptr<Sink> make_sink(std::string_view base_uri,
+                                    std::string_view path);
 
     /// @brief Create a collection of data sinks, either chunk or shard.
     /// @param[in] base_uri The base URI for the sinks.
@@ -36,15 +44,14 @@ struct SinkCreator final
     /// @brief Create a collection of metadata sinks for a Zarr dataset.
     /// @param[in] version The Zarr version.
     /// @param[in] base_uri The base URI for the dataset.
-    /// @param[in] n_arrays The number of data arrays.
-    /// @param[out] metadata_sinks The sinks created.
+    /// @param[out] metadata_sinks The sinks created, keyed by path.
+    /// @return True iff all metadata sinks were created successfully.
     /// @throws std::runtime_error if @p base_uri is not valid, or if, for S3
     ///         sinks, the bucket does not exist.
     [[nodiscard]] bool make_metadata_sinks(
       ZarrVersion version,
       const std::string& base_uri,
-      size_t n_arrays,
-      std::vector<std::unique_ptr<Sink>>& metadata_sinks);
+      std::unordered_map<std::string, std::unique_ptr<Sink>>& metadata_sinks);
 
   private:
     std::shared_ptr<common::ThreadPool> thread_pool_;
@@ -63,16 +70,14 @@ struct SinkCreator final
     [[nodiscard]] bool make_files_(std::queue<std::string>& file_paths,
                                    std::vector<std::unique_ptr<Sink>>& sinks);
 
-    /// @brief Create a collection of metadata sinks.
-    /// @param[in] base_uri The base URI for the sinks.
-    /// @param[in] dir_paths The directories to create.
-    /// @param[in] file_paths The files to create.
-    /// @param[out] metadata_sinks The sinks created.
-    [[nodiscard]] bool make_metadata_sinks_(
-      const std::string& base_uri,
-      std::queue<std::string>& dir_paths,
-      std::queue<std::string>& file_paths,
-      std::vector<std::unique_ptr<Sink>>& metadata_sinks);
+    /// @brief Parallel create a collection of files, keyed by path.
+    /// @param[in] base_dir The base directory for the files.
+    /// @param[in] file_paths Paths to the files to create, relative to @p base_dir.
+    /// @param[out] sinks The sinks created, keyed by path.
+    [[nodiscard]] bool make_files_(
+      const std::string& base_dir,
+      const std::vector<std::string>& file_paths,
+      std::unordered_map<std::string, std::unique_ptr<Sink>>& sinks);
 
     /// @brief Check whether an S3 bucket exists.
     /// @param[in] bucket_name The name of the bucket to check.
@@ -87,5 +92,10 @@ struct SinkCreator final
       std::string_view bucket_name,
       std::queue<std::string>& object_keys,
       std::vector<std::unique_ptr<Sink>>& sinks);
+
+    [[nodiscard]] bool make_s3_objects_(
+      std::string_view bucket_name,
+      std::vector<std::string>& object_keys,
+      std::unordered_map<std::string, std::unique_ptr<Sink>>& sinks);
 };
 } // namespace acquire::sink::zarr