+++ src/zarr.v3.hh
@@ -27,7 +27,6 @@ struct ZarrV3 final : public Zarr
 
     // mutable metadata, changes on flush
     void write_group_metadata_() const override;
-    void write_array_metadata_(size_t level) const override;
 };
 } // namespace acquire::sink::zarr
 #endif // H_ACQUIRE_STORAGE_ZARR_V3_V0