+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Calling `acquire_get_configuration` with a Zarr storage device now returns a URI of the storage device, with file://
   scheme indicator and absolute path, assuming localhost.
+- Array metadata is updated every time data is flushed to disk. 
 
 ## [0.1.11](https://github.com/acquire-project/acquire-driver-zarr/compare/v0.1.10..v0.1.11) - 2024-04-22